+++ examples/simple_react/amplify/backend.ts
@@ -1,8 +1,0 @@
-import { defineBackend } from '@aws-amplify/backend';
-import { auth } from './auth/resource.js';
-import { data } from './data/resource.js';
-
-defineBackend({
-  auth,
-  data,
-});