+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @aws-amplify/backend-function
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [cb855dfa5]
+  - @aws-amplify/backend-output-storage@0.2.4
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [70685f36b]
+  - @aws-amplify/backend-output-storage@0.2.3
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [65fe3a8fd]
+- Updated dependencies [cd5feeed0]
+  - @aws-amplify/plugin-types@0.4.1
+  - @aws-amplify/backend-output-storage@0.2.2
+
 ## 0.2.0
 
 ### Minor Changes