+++ packages/backend-output-storage/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @aws-amplify/backend-output-storage
 
+## 0.2.4
+
+### Patch Changes
+
+- cb855dfa5: chore: refactor packageJsonReader and generate installationIds from hostname
+- Updated dependencies [cb855dfa5]
+  - @aws-amplify/platform-core@0.2.2
+
+## 0.2.3
+
+### Patch Changes
+
+- 70685f36b: Add usage data metrics
+- Updated dependencies [70685f36b]
+  - @aws-amplify/platform-core@0.2.1
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [07b0dfc9f]
+  - @aws-amplify/backend-output-schemas@0.4.0
+
 ## 0.2.1
 
 ### Patch Changes